--- keyboards/drakon/keymaps/logo/readme.md
@@ -1,1 +1,0 @@
-# The logo keymap for drakon
